# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface TravelRecommendationRequest {
  /** City or region to get recommendations for */
  destination: string;
  /** Travel vibe tags (e.g. foodie, nightlife, culture) */
  vibes?: string[];
}

export interface TravelBusiness {
  name: string;
  category: string;
  description: string;
  neighborhood: string;
  mustTry: string;
}

export interface TravelNeighborhood {
  name: string;
  vibe: string;
  highlights: string[];
  safetyNote: string;
}

export interface TravelEvent {
  name: string;
  type: string;
  description: string;
  timing: string;
}

export interface TravelRecommendations {
  destination: string;
  summary: string;
  businesses: TravelBusiness[];
  neighborhoods: TravelNeighborhood[];
  events: TravelEvent[];
  safetyTips: string[];
  localInsights: string[];
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export type OpenaiImageInputSize = typeof OpenaiImageInputSize[keyof typeof OpenaiImageInputSize];


export const OpenaiImageInputSize = {
  '1024x1024': '1024x1024',
  '512x512': '512x512',
  '256x256': '256x256',
} as const;

export interface OpenaiImageInput {
  prompt: string;
  size?: OpenaiImageInputSize;
}

export interface OpenaiImageOutput {
  b64_json: string;
}

export interface OpenaiError {
  error: string;
}