# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /travel/recommendations:
    post:
      operationId: getTravelRecommendations
      tags: [travel]
      summary: Get AI travel recommendations for a destination
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TravelRecommendationRequest"
      responses:
        "200":
          description: AI-generated travel recommendations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelRecommendations"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    TravelRecommendationRequest:
      type: object
      properties:
        destination:
          type: string
          description: City or region to get recommendations for
        vibes:
          type: array
          items:
            type: string
          description: Travel vibe tags (e.g. foodie, nightlife, culture)
      required:
        - destination
    TravelRecommendations:
      type: object
      properties:
        destination:
          type: string
        summary:
          type: string
        businesses:
          type: array
          items:
            $ref: "#/components/schemas/TravelBusiness"
        neighborhoods:
          type: array
          items:
            $ref: "#/components/schemas/TravelNeighborhood"
        events:
          type: array
          items:
            $ref: "#/components/schemas/TravelEvent"
        safetyTips:
          type: array
          items:
            type: string
        localInsights:
          type: array
          items:
            type: string
      required:
        - destination
        - summary
        - businesses
        - neighborhoods
        - events
        - safetyTips
        - localInsights
    TravelBusiness:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        neighborhood:
          type: string
        mustTry:
          type: string
      required:
        - name
        - category
        - description
        - neighborhood
        - mustTry
    TravelNeighborhood:
      type: object
      properties:
        name:
          type: string
        vibe:
          type: string
        highlights:
          type: array
          items:
            type: string
        safetyNote:
          type: string
      required:
        - name
        - vibe
        - highlights
        - safetyNote
    TravelEvent:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        description:
          type: string
        timing:
          type: string
      required:
        - name
        - type
        - description
        - timing
...