# lib/api-zod/src/generated/types/travelEvent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TravelEvent {
  name: string;
  type: string;
  description: string;
  timing: string;
}