# lib/api-zod/src/generated/types/travelNeighborhood.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TravelNeighborhood {
  name: string;
  vibe: string;
  highlights: string[];
  safetyNote: string;
}