# lib/api-zod/src/generated/types/travelRecommendationRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TravelRecommendationRequest {
  /** City or region to get recommendations for */
  destination: string;
  /** Travel vibe tags (e.g. foodie, nightlife, culture) */
  vibes?: string[];
}