# lib/api-zod/src/generated/types/travelRecommendations.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TravelBusiness } from './travelBusiness';
import type { TravelEvent } from './travelEvent';
import type { TravelNeighborhood } from './travelNeighborhood';

export interface TravelRecommendations {
  destination: string;
  summary: string;
  businesses: TravelBusiness[];
  neighborhoods: TravelNeighborhood[];
  events: TravelEvent[];
  safetyTips: string[];
  localInsights: string[];
}